ANALYTICS-001 fix repository PK and OrderBy

+++ backend/src/main/java/com/bioinformatics/dashboard/analytics/repository/EvidenceDistributionRepository.java
@@ -3,5 +3,5 @@ package com.bioinformatics.dashboard.analytics.repository;
 import com.bioinformatics.dashboard.analytics.entity.EvidenceDistribution;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EvidenceDistributionRepository extends JpaRepository<EvidenceDistribution, Integer> {
+public interface EvidenceDistributionRepository extends JpaRepository<EvidenceDistribution, Short> {
 }
